updated administration service to metadataservice
