+++ protos/dcs/dcs.proto
@@ -2,7 +2,6 @@ syntax = "proto3";
 
 package dcs;
 
-import "dcs/administration/v0/administration.proto";
 import "dcs/atmosphere/v0/atmosphere.proto";
 import "dcs/coalition/v0/coalition.proto";
 import "dcs/common/v0/common.proto";
@@ -10,6 +9,7 @@ import "dcs/controller/v0/controller.proto";
 import "dcs/custom/v0/custom.proto";
 import "dcs/group/v0/group.proto";
 import "dcs/hook/v0/hook.proto";
+import "dcs/metadata/v0/metadata.proto";
 import "dcs/mission/v0/mission.proto";
 import "dcs/net/v0/net.proto";
 import "dcs/srs/v0/srs.proto";
