--- stubs/src/administration.rs
@@ -1,3 +1,0 @@
-pub mod v0 {
-    tonic::include_proto!("dcs.administration.v0");
-}
