+++ stubs/src/lib.rs
@@ -3,7 +3,6 @@
 #![allow(clippy::derive_partial_eq_without_eq)]
 #![allow(clippy::large_enum_variant)]
 
-pub mod administration;
 pub mod atmosphere;
 pub mod coalition;
 pub mod common;
@@ -11,6 +10,7 @@ pub mod controller;
 pub mod custom;
 pub mod group;
 pub mod hook;
+pub mod metadata;
 pub mod mission;
 pub mod net;
 pub mod srs;
