+++ stubs/src/metadata.rs
@@ -1,0 +1,3 @@
+pub mod v0 {
+    tonic::include_proto!("dcs.metadata.v0");
+}
